+++ index.html
@@ -10,7 +10,7 @@
 </head>
 
 <body>
-  <script src="main.js"></script>
+  <script src="main.js" type = "module"></script>
   <main>
 
     <!-- Фильтрация изображений от других пользователей -->
